requirements.txt
aiohttp==3.11.13
aiosignal==1.3.2
asyncache==0.3.1
attrs==25.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
pipdeptree==2.25.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.3.0
yarl==1.18.3
